Improve folder-structure for MentorsPage
