--- src/features/MentorPage/MentorList/MentorCard/Expanded/index.ts
@@ -1,3 +1,0 @@
-import { MentorCard } from './MentorCard';
-
-export default MentorCard;
