--- src/features/MentorPage/MentorList/MentorCard/List/index.ts
@@ -1,3 +1,0 @@
-import { ListCard } from './ListCard';
-
-export default ListCard;
