--- src/features/MentorPage/MentorList/index.ts
@@ -1,3 +1,0 @@
-import MentorList from './MentorList';
-
-export default MentorList;
