+++ src/features/MentorPage/components/MentorList/MentorCard/Expanded/Skills.tsx
@@ -1,8 +1,8 @@
+import { useTranslation } from 'react-i18next';
+
 import styled from 'styled-components';
 import { Text } from '@/components/Text/Text';
 import { SimpleChip } from '@/components/Chip';
-
-import { useTranslation } from 'react-i18next';
 
 type Props = {
   skills: Array<string>;
