+++ src/features/MentorPage/components/MentorList/MentorCard/List/Languages.tsx
@@ -1,17 +1,16 @@
-import styled from 'styled-components';
 import { useTranslation } from 'react-i18next';
 
+import styled from 'styled-components';
 import { SimpleChip } from '@/components/Chip';
-import { palette } from '@/components/variables';
-import Text from '@/components/Text';
+import { Text } from '@/components/Text/Text';
 
-export const Skills = ({ skills }: { skills: Array<string> }) => {
+export const Languages = ({ languages }: { languages: Array<string> }) => {
   const { t } = useTranslation('mentors');
   return (
     <Container>
-      <Header variant="h3">{t('card.skills')}</Header>
+      <Header variant="h3">{t('card.languages')}</Header>
       <Chips>
-        {skills.map(item => (
+        {languages.map(item => (
           <SimpleChip key={item} text={item} />
         ))}
       </Chips>
@@ -22,9 +21,7 @@ export const Skills = ({ skills }: { skills: Array<string> }) => {
 const Container = styled.div`
   display: flex;
   flex-direction: column;
-  gap: 1rem;
-  height: fit-content;
-  max-height: 8.6rem;
+  gap: 0.5rem;
 `;
 
 const Header = styled(Text)`
@@ -36,18 +33,10 @@ const Chips = styled.div`
   flex: 0 0 auto;
   flex-wrap: wrap;
   gap: 0.5rem;
-  height: 6rem;
+  height: fit-content;
   justify-content: flex-start;
+  max-height: 7rem;
   overflow: hidden;
   position: relative;
   width: 100%;
-  &:after {
-    background: linear-gradient(transparent, ${palette.white});
-    content: '';
-    display: block;
-    height: 2.5rem;
-    position: absolute;
-    top: 3.5rem;
-    width: 100%;
-  }
 `;
