+++ src/features/MentorPage/components/MentorsFilter/index.tsx
@@ -1,11 +1,12 @@
 import { useState } from 'react';
+
+import { selectSkills } from '../../mentorPageApi';
+import { useAppSelector } from '@/store';
+import { useTranslation } from 'react-i18next';
+
 import Filters from './Filters/Filters';
 import MentorSearch from './MentorSearch';
 import SkillChips from './Skills/';
-
-import { selectSkills } from '../mentorPageApi';
-import { useAppSelector } from '@/store';
-import { useTranslation } from 'react-i18next';
 
 const MentorsFilter = () => {
   const skills = useAppSelector(selectSkills());
