--- src/features/MentorPage/index.ts
@@ -1,3 +1,0 @@
-import MentorPage from './MentorPage';
-
-export default MentorPage;
