+++ src/store.ts
@@ -5,7 +5,7 @@ import {
 } from '@reduxjs/toolkit';
 import { TypedUseSelectorHook, useSelector, useDispatch } from 'react-redux';
 import { mentorsApi } from '@/features/MentorPage/mentorPageApi';
-import { mentorsFilter } from '@/features/MentorPage/MentorsFilter/mentorsFilterSlice';
+import { mentorsFilter } from '@/features/MentorPage/mentorsFilterSlice';
 
 const rootReducer = combineReducers({
   [mentorsApi.reducerPath]: mentorsApi.reducer,
